(maint) Use ModuleLoader in Environment#deploy
Previously, we continued to use the Puppetfile to sync content when
doing a deploy via an Environment. This patch moves us to use the
ModuleLoader::Puppetfile in its place.

To do so, we change the ModuleLoader::Puppetfile#load to be `load!` and
provide a non-raising version for the Environment#deploy to call. This
is because the Environment depends on the the Loader not raising during
invocation so it may do further processing, while other invocations
require the Loader to raise.

We also now pass the pool_size through the overrides to the environment
as previously the Puppetfile class the setting set on it globally.

We also extend the mock environment to be purgeable in line with the
actual environments it mocks.