(maint) Do not use reference/alternates with shellgit, pt. 1
Rugged does not seem to have issues with alternates/reference repo(s)
but shellgit will quickly fail if it encounters any issue with the
object db(s) refrenced by alternates (as well as printing a lot of
warning about bad paths and objects).

I _think_ this has been exacerbated by our more thorough usage of
`--prune` when fetching and a more performant fix may be through
tweaking our pruning while fetching. Or with newer git versions
potentially using multiple worktrees.

However, this approach definitely resolves the issue see by some
customers (if using a too-large of hammer). This work is done solely for
the purposes of a hotfix while we investigate the other approaches.

This will cause all "thin" repositories to no longer be "thin" and
consume same amount of disk space as a control repo.